Restructuring of the Id/Key/Owner implementation used for all collection keys. This replaces the use of the term Id in most cases with Key. Id is the lower bits of Key only. Owner is now always a value from 1 to 255 (zero if unknown/unset).

+++ Nova/WinForms/Gui/Controls/SelectionDetail.cs
@@ -61,7 +61,7 @@ namespace Nova.WinForms.Gui
         /// Initializes a new instance of the SelectionDetail class.
         /// </Summary>
         public SelectionDetail(EmpireData empireState,
-                               List<int> deletedFleets,
+                               List<long> deletedFleets,
                                ClientState stateData)
         {
             this.empireState = empireState;
